+++ src/test/java/de/chrgroth/jsonstore/store/model/TestDataVersion2.java
@@ -1,0 +1,7 @@
+package de.chrgroth.jsonstore.store.model;
+
+public class TestDataVersion2 {
+	public int id;
+	public String name;
+	public String description;
+}